Move shared test code to test library.
